Refactor: Banner React Query사용으로 SSR로 수정
