+++ app/(route)/main/_components/Body/index.tsx
@@ -1,0 +1,24 @@
+'use client';
+
+import { useState } from 'react';
+
+import AllActivityLists from '../AllActivityLists';
+import HotActivityLists from '../HotActivityLists';
+import SearchBox from '../SearchBox';
+
+export default function Body() {
+  const [searchValue, setSearchValue] = useState('');
+  const handleSearch = (value: string) => {
+    setSearchValue(value);
+  };
+
+  return (
+    <>
+      <SearchBox onSearch={handleSearch} />
+      <div className="relative bottom-[20px] mx-auto w-full px-[16px] lg:max-w-[1200px] mobile:bottom-[20px] mobile:px-[24px]">
+        {searchValue === '' && <HotActivityLists />}
+        <AllActivityLists searchValue={searchValue} />
+      </div>
+    </>
+  );
+}
